allow deletion from custom lists

+++ Sect28_ToDoList_V2/app.js
@@ -4,6 +4,7 @@ app.use(express.json());
 app.use(express.urlencoded({extended : true}))
 app.use(express.static("public"));
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 mongoose.connect("mongodb://127.0.0.1:27017/todoListDB");
 
@@ -13,19 +14,25 @@ const itemsSchema = new mongoose.Schema({
 
 const Item = mongoose.model("Item", itemsSchema);
 
-const workout = new Item({
-    name: "Workout"
+const item1 = new Item({
+    name: "Welcome to your ToDo List!"
 });
 
-const leetcode = new Item({
-    name: "Leetcode"
+const item2 = new Item({
+    name: "Press the + button to add an item."
 });
 
-const udemy = new Item({
-    name: "Udemy"
+const item3 = new Item({
+    name: "<-- Check the box to remove an item."
 });
 
-const defaultItems = [workout, leetcode, udemy];
+const defaultItems = [item1, item2, item3];
+
+const listSchema = new mongoose.Schema({
+    name: String,
+    items: [itemsSchema]
+});
+const List = mongoose.model("List", listSchema)
 
 
 
@@ -49,7 +56,7 @@ app.get("/", function(req, res) {
         } else {
             res.render("list", {
                 listTitle : "Today",
-                newListItems : items
+                listItems : items
             });
         }
         
@@ -60,36 +67,90 @@ app.get("/", function(req, res) {
     
 });
 
+app.get("/:customListName", function(req, res) {
+    const customListName = _.capitalize(req.params.customListName);
+
+    List.findOne({name: customListName})
+    .then((foundList) => {
+        if(!foundList) {
+            console.log("list not found");
+            const list = new List({
+                name: customListName,
+                items: defaultItems
+            });
+            list.save()
+            .then(() => {
+                res.redirect("/"+customListName);
+            });
+        } else {
+            console.log("list found");
+            res.render("list", {
+                listTitle: foundList.name,
+                listItems: foundList.items
+            });
+        }
+    })
+    .catch((err) => {
+        console.log(err);
+    })
+
+
+});
+
 app.post("/", function(req, res) {
 
     const itemName = req.body.newItem;
+    const listName = req.body.list;
     const newItem = new Item({
         name : itemName
     });
-    newItem.save();
 
-    res.redirect("/");
+    if(listName === "Today") {
+        newItem.save();
+        res.redirect("/");
+
+    } else {
+        List.findOne({name: listName})
+        .then((foundList) => {
+            foundList.items.push(newItem);
+            foundList.save();
+            res.redirect("/" + listName);
+        })
+    }
+    
+
+    
 });
 
 app.post("/delete", function(req, res) {
     const checkedItemID = req.body.checkbox;
-    Item.deleteOne({_id: checkedItemID})
-    .then(() => {
-        console.log("Successfully deleted checked item from database.");
-        res.redirect("/");
-    })
-    .catch((err) => {
-        console.log(err);
-    });
+    const listName = req.body.listName;
+
+    if(listName === "Today") {
+        Item.deleteOne({_id: checkedItemID})
+        .then(() => {
+            console.log("Successfully deleted checked item from database.");
+            res.redirect("/");
+        })
+        .catch((err) => {
+            console.log(err);
+        });
+
+    } else {
+        List.findOneAndUpdate({name: listName}, {
+            $pull: {items: {_id: checkedItemID}}
+        })
+        .then(() => {
+            res.redirect("/" + listName);
+        })
+        .catch((err) => {
+            console.log(err);
+        });
+    }
+
+    
 });
 
-
-app.get("/work", function(req, res){
-    res.render("list", {
-        listTitle : "Work List",
-        newListItems : workItems
-    })
-});
 
 app.get("/about", function(req, res) {
     res.render("about");
